Omit plugin name only when vl.Plugin == vl.Type
This is a breaking change, as it affects how the metric names are
generated.

Before this change, metric with vl.Plugin "ping" and vl.Type
"ping_droprate" was named as "ping_droprate".

After this change, it would be named as "ping_ping_droprate", which works around
issue #22 where different metrics (from one plugin) were given the same name
due to previous naming scheme.